+++ CHANGELOG.rst
@@ -17,9 +17,17 @@ Unreleased
 ----------
 * Nothing
 
-[3.28.21]
+[3.28.23]
 ---------
 * Fix cornerstone character limit bug with dict database table
+
+[3.28.22]
+---------
+* fix: Adding error handling for role assignment backfill management command
+
+[3.28.21]
+---------
+* bug: The exporter now properly handles instances when enterprise customer catalogs do no need updates.
 
 [3.28.20]
 ---------